+++ Libs/TemplateLoader.php
@@ -39,7 +39,6 @@ class TemplateLoader {
         // Add a filter to the save post to inject out template into the page cache
         \add_filter('wp_insert_post_data', [$this, 'registerProjectTemplates']);
 
-
         // Add a filter to the template include to determine if the page has our
         // template assigned and return it's path
         \add_filter('template_include', [$this, 'viewProjectTemplate']);
@@ -53,13 +52,13 @@ class TemplateLoader {
     /**
      * Adds our template to the page dropdown for v4.7+
      *
-     * @param array $posts_templates
+     * @param array $postsTemplates
      * @return array
      */
-    public function addNewTemplate($posts_templates) {
-        $posts_templates = \array_merge($posts_templates, $this->templates);
+    public function addNewTemplate($postsTemplates) {
+        $postsTemplates = \array_merge($postsTemplates, $this->templates);
 
-        return $posts_templates;
+        return $postsTemplates;
     }
 
     /**
@@ -71,7 +70,7 @@ class TemplateLoader {
      */
     public function registerProjectTemplates($atts) {
         // Create the key used for the themes cache
-        $cache_key = 'page_templates-' . \md5(\get_theme_root() . '/' . \get_stylesheet());
+        $cacheKey = 'page_templates-' . \md5(\get_theme_root() . '/' . \get_stylesheet());
 
         // Retrieve the cache list.
         // If it doesn't exist, or it's empty prepare an array
@@ -82,7 +81,7 @@ class TemplateLoader {
         }
 
         // New cache, therefore remove the old one
-        \wp_cache_delete($cache_key, 'themes');
+        \wp_cache_delete($cacheKey, 'themes');
 
         // Now add our template to the list of templates by merging our templates
         // with the existing templates array from the cache.
@@ -90,7 +89,7 @@ class TemplateLoader {
 
         // Add the modified cache to allow WordPress to pick it up for listing
         // available templates
-        \wp_cache_add($cache_key, $templates, 'themes', 1800);
+        \wp_cache_add($cacheKey, $templates, 'themes', 1800);
 
         return $atts;
     }