+++ templates/partials/alliance/alliance-participation.php
@@ -3,7 +3,7 @@
 if(!empty($allianceParticipation)) {
     ?>
     <div class="table-responsive table-local-scan table-local-scan-alliances table-eve-intel">
-        <table class="table table-condensed table-sortable" data-haspaging="no" data-order='[[ 1, "desc" ]]'>
+        <table class="table table-condensed table-sortable eve-intel-alliance-participation-list" data-haspaging="no" data-order='[[ 1, "desc" ]]'>
             <thead>
                 <th><?php echo \__('Alliance Name', 'eve-online-intel-tool'); ?></th>
                 <th><?php echo \__('Count', 'eve-online-intel-tool'); ?></th>
@@ -12,15 +12,17 @@ if(!empty($allianceParticipation)) {
             foreach($allianceParticipation as $allianceList) {
                 foreach($allianceList as $alliance) {
                     ?>
-                    <tr data-highlight="alliance-<?php echo $alliance['allianceID']; ?>">
+                    <tr class="eve-intel-alliance-participation-item eve-intel-alliance-id-<?php echo $alliance['allianceID']; ?>" data-highlight="alliance-<?php echo $alliance['allianceID']; ?>">
                         <td>
                             <?php
-                            \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('data/alliance-logo', [
+                            \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('partials/alliance/alliance-logo', [
                                 'data' => $alliance,
                                 'pluginSettings' => $pluginSettings
                             ]);
 
-                            echo $alliance['allianceName'];
+                            \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('partials/alliance/alliance-information', [
+                                'data' => $alliance
+                            ]);
                             ?>
                         </td>
                         <td class="table-data-count">