+++ templates/partials/dscan/dscan-breakdown.php
@@ -4,7 +4,7 @@
     -->
     <div class="col-md-4 col-sm-6 col-lg-3">
         <?php
-        \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('data/ship-classes',[
+        \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('partials/ship/ship-classes', [
             'title' => \__('All', 'eve-online-intel-tool'),
             'count' => $countAll,
             'shipClassList' => (!empty($intelData['dscanDataAll']['data'])) ? $intelData['dscanDataAll']['data'] : null,
@@ -18,7 +18,7 @@
     -->
     <div class="col-md-4 col-sm-6 col-lg-3">
         <?php
-        \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('data/ship-classes',[
+        \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('partials/ship/ship-classes', [
             'title' => \__('On Grid', 'eve-online-intel-tool'),
             'count' => $countOnGrid,
             'shipClassList' => (!empty($intelData['dscanDataOnGrid']['data'])) ? $intelData['dscanDataOnGrid']['data'] : null,
@@ -32,7 +32,7 @@
     -->
     <div class="col-md-4 col-sm-6 col-lg-3">
         <?php
-        \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('data/ship-classes',[
+        \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('partials/ship/ship-classes', [
             'title' => \__('Off Grid', 'eve-online-intel-tool'),
             'count' => $countOffGrid,
             'shipClassList' => (!empty($intelData['dscanDataOffGrid']['data'])) ? $intelData['dscanDataOffGrid']['data'] : null,
@@ -46,7 +46,7 @@
     -->
     <div class="col-md-4 col-sm-6 col-lg-3">
         <?php
-        \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('data/ship-types',[
+        \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('partials/ship/ship-types', [
             'title' => \__('Ship Types', 'eve-online-intel-tool'),
             'shipTypeList' => $intelData['dscanDataShipTypes'],
             'pluginSettings' => $pluginSettings