+++ templates/partials/fleet/fleet-composition.php
@@ -15,7 +15,7 @@ if(\is_array($fleetOverview) && \count($fleetOverview)) {
                 <tr data-highlight="shiptype-<?php echo \sanitize_title($data['shipType']); ?>">
                     <td>
                         <?php
-                        \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('data/pilot-avatar', [
+                        \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('partials/pilot/pilot-avatar', [
                             'data' => [
                                 'characterID' => $data['pilotID'],
                                 'characterName' => $data['pilotName']
@@ -23,7 +23,12 @@ if(\is_array($fleetOverview) && \count($fleetOverview)) {
                             'pluginSettings' => $pluginSettings
                         ]);
 
-                        echo $data['pilotName'];
+                        \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('partials/pilot/pilot-information', [
+                            'data' => [
+                                'characterID' => $data['pilotID'],
+                                'characterName' => $data['pilotName']
+                            ]
+                        ]);
                         ?>
                     </td>
                     <td>