+++ templates/single-fleetcomposition.php
@@ -3,16 +3,16 @@
 ?>
 
 <header class="page-title">
-    <h1>
-        <?php echo \__('Fleet Composition', 'eve-online-intel-tool'); ?>
-    </h1>
+    <h1><?php echo \__('Fleet Composition', 'eve-online-intel-tool'); ?></h1>
 
     <?php
     if(!empty($intelData['eveTime'])) {
-        echo '<small>' . \__('EVE Time:', 'eve-online-intel-tool') . ' ' . $intelData['eveTime'] . '</small>';
+        ?>
+        <div><p><small><?php echo \__('EVE Time:', 'eve-online-intel-tool') . ' ' . $intelData['eveTime']; ?></small></p></div>
+        <?php
     }
 
-    \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('extra/buttons');
+    \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('partials/extra/buttons');
     ?>
 </header>
 
@@ -25,7 +25,7 @@
                 -->
                 <?php
                 if(\is_array($intelData['shipTypes']) && \count($intelData['shipTypes'])) {
-                    \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('data/fleet-information',[
+                    \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('partials/fleet/fleet-information', [
                         'pilotCount' => \count($intelData['pilotList']),
                         'fleetInformation' => $intelData['fleetInformation'],
                         'pluginSettings' => $pluginSettings
@@ -41,7 +41,7 @@
                 <div class="col-md-6 col-lg-3">
                     <?php
                     if(\is_array($intelData['shipClasses']) && \count($intelData['shipClasses'])) {
-                        \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('data/ship-classes',[
+                        \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('partials/ship/ship-classes', [
                             'title' => \__('Ship Classes', 'eve-online-intel-tool'),
                             'shipClassList' => $intelData['shipClasses'],
                             'pluginSettings' => $pluginSettings
@@ -56,7 +56,7 @@
                 <div class="col-md-6 col-lg-3">
                     <?php
                     if(\is_array($intelData['shipTypes']) && \count($intelData['shipTypes'])) {
-                        \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('data/ship-types',[
+                        \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('partials/ship/ship-types', [
                             'title' => \__('Ship Types', 'eve-online-intel-tool'),
                             'shipTypeList' => $intelData['shipTypes'],
                             'pluginSettings' => $pluginSettings
@@ -70,7 +70,7 @@
                 -->
                 <div class="col-md-12 col-lg-6">
                     <?php
-                    \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('data/fleet-composition',[
+                    \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('partials/fleet/fleet-composition', [
                         'fleetOverview' => $intelData['fleetOverview'],
                         'pluginSettings' => $pluginSettings
                     ]);
@@ -79,12 +79,20 @@
             </div>
 
             <div class="fleetcomposition-result row">
+                <div class="clearfix">
+                    <div class="col-sm-6 col-md-4 col-lg-3 clearfix">
+                        <header class="eve-intel-section-header">
+                            <h3><?php echo \__('Participation Details', 'eve-online-intel-tool'); ?></h3>
+                        </header>
+                    </div>
+                </div>
+
                 <!--
                 // Alliance participation
                 -->
                 <div class="col-md-6 col-lg-3">
                     <?php
-                    \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('data/alliance-participation',[
+                    \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('partials/alliance/alliance-participation', [
                         'allianceCount' => \count($intelData['allianceList']),
                         'allianceParticipation' => $intelData['allianceParticipation'],
                         'pluginSettings' => $pluginSettings
@@ -97,7 +105,7 @@
                 -->
                 <div class="col-md-6 col-lg-3">
                     <?php
-                    \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('data/corporation-participation',[
+                    \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('partials/corporation/corporation-participation', [
                         'corporationCount' => \count($intelData['corporationList']),
                         'corporationParticipation' => $intelData['corporationParticipation'],
                         'pluginSettings' => $pluginSettings
@@ -110,7 +118,7 @@
                 -->
                 <div class="col-md-12 col-lg-6">
                     <?php
-                    \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('data/pilot-participation',[
+                    \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\TemplateHelper::getTemplate('partials/pilot/pilot-participation', [
                         'pilotCount' => \count($intelData['pilotList']),
                         'pilotParticipation' => $intelData['pilotParticipation'],
                         'pluginSettings' => $pluginSettings